move ctlogs at pkg/
